Revert part of "Updated folder structure to reduce code duplication between collector and targetallocator & Minor changes"
This reverts commit 367c6237b9c0e9c77d141229748d91429ff31791 - only the folder structure change
