+++ pkg/collector/reconcile/configmap.go
@@ -19,7 +19,6 @@ import (
 	"fmt"
 	"reflect"
 
-	"gopkg.in/yaml.v2"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -29,15 +28,14 @@ import (
 
 	"github.com/open-telemetry/opentelemetry-operator/pkg/collector"
 	"github.com/open-telemetry/opentelemetry-operator/pkg/naming"
-	"github.com/open-telemetry/opentelemetry-operator/pkg/targetallocator"
 )
 
 // +kubebuilder:rbac:groups="",resources=configmaps,verbs=get;list;watch;create;update;patch;delete
 
-// CollectorConfigMaps reconciles the config map(s) required for the collector instance in the current context.
-func CollectorConfigMaps(ctx context.Context, params Params) error {
+// ConfigMaps reconciles the config map(s) required for the instance in the current context.
+func ConfigMaps(ctx context.Context, params Params) error {
 	desired := []corev1.ConfigMap{
-		desiredCollectorConfigMap(ctx, params),
+		desiredConfigMap(ctx, params),
 	}
 
 	// first, handle the create/update parts
@@ -46,53 +44,14 @@ func CollectorConfigMaps(ctx context.Context, params Params) error {
 	}
 
 	// then, delete the extra objects
-	opts := []client.ListOption{
-		client.InNamespace(params.Instance.Namespace),
-		client.MatchingLabels(map[string]string{
-			"app.kubernetes.io/instance":   fmt.Sprintf("%s.%s", params.Instance.Namespace, params.Instance.Name),
-			"app.kubernetes.io/managed-by": "opentelemetry-operator",
-		}),
-	}
-	if err := deleteConfigMaps(ctx, params, desired, opts); err != nil {
+	if err := deleteConfigMaps(ctx, params, desired); err != nil {
 		return fmt.Errorf("failed to reconcile the configmaps to be deleted: %v", err)
 	}
 
 	return nil
 }
 
-// TAConfigMaps reconciles the config map(s) required for the target allocator instance in the current context.
-func TAConfigMaps(ctx context.Context, params Params) error {
-	desired := []corev1.ConfigMap{}
-
-	if params.Instance.Spec.TargetAllocator.Enabled {
-		cm, err := desiredTAConfigMap(ctx, params)
-		if err != nil {
-			return fmt.Errorf("failed to parse config: %v", err)
-		}
-		desired = append(desired, cm)
-	}
-
-	// first, handle the create/update parts
-	if err := expectedConfigMaps(ctx, params, desired, true); err != nil {
-		return fmt.Errorf("failed to reconcile the expected configmaps: %v", err)
-	}
-
-	// then, delete the extra objects
-	opts := []client.ListOption{
-		client.InNamespace(params.Instance.Namespace),
-		client.MatchingLabels(map[string]string{
-			"app.kubernetes.io/instance":   fmt.Sprintf("%s.%s", params.Instance.Name, "targetallocator"),
-			"app.kubernetes.io/managed-by": "opentelemetry-operator",
-		}),
-	}
-	if err := deleteConfigMaps(ctx, params, desired, opts); err != nil {
-		return fmt.Errorf("failed to reconcile the configmaps to be deleted: %v", err)
-	}
-
-	return nil
-}
-
-func desiredCollectorConfigMap(_ context.Context, params Params) corev1.ConfigMap {
+func desiredConfigMap(_ context.Context, params Params) corev1.ConfigMap {
 	name := naming.ConfigMap(params.Instance)
 	labels := collector.Labels(params.Instance)
 	labels["app.kubernetes.io/name"] = name
@@ -108,40 +67,6 @@ func desiredCollectorConfigMap(_ context.Context, params Params) corev1.ConfigMa
 			"collector.yaml": params.Instance.Spec.Config,
 		},
 	}
-}
-
-func desiredTAConfigMap(_ context.Context, params Params) (corev1.ConfigMap, error) {
-	name := naming.TAConfigMap(params.Instance)
-	labels := targetallocator.Labels(params.Instance)
-	labels["app.kubernetes.io/name"] = name
-
-	promConfig, err := GetPromConfig(params)
-	if err != nil {
-		return corev1.ConfigMap{}, err
-	}
-
-	taConfig := make(map[interface{}]interface{})
-	taConfig["label_selector"] = map[string]string{
-		"app.kubernetes.io/instance":   fmt.Sprintf("%s.%s", params.Instance.Namespace, params.Instance.Name),
-		"app.kubernetes.io/managed-by": "opentelemetry-operator",
-	}
-	taConfig["config"] = promConfig
-	taConfigYAML, err := yaml.Marshal(taConfig)
-	if err != nil {
-		return corev1.ConfigMap{}, err
-	}
-
-	return corev1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        name,
-			Namespace:   params.Instance.Namespace,
-			Labels:      labels,
-			Annotations: params.Instance.Annotations,
-		},
-		Data: map[string]string{
-			"targetallocator.yaml": string(taConfigYAML),
-		},
-	}, nil
 }
 
 func expectedConfigMaps(ctx context.Context, params Params, expected []corev1.ConfigMap, retry bool) error {
@@ -210,7 +135,14 @@ func expectedConfigMaps(ctx context.Context, params Params, expected []corev1.Co
 	return nil
 }
 
-func deleteConfigMaps(ctx context.Context, params Params, expected []corev1.ConfigMap, opts []client.ListOption) error {
+func deleteConfigMaps(ctx context.Context, params Params, expected []corev1.ConfigMap) error {
+	opts := []client.ListOption{
+		client.InNamespace(params.Instance.Namespace),
+		client.MatchingLabels(map[string]string{
+			"app.kubernetes.io/instance":   fmt.Sprintf("%s.%s", params.Instance.Namespace, params.Instance.Name),
+			"app.kubernetes.io/managed-by": "opentelemetry-operator",
+		}),
+	}
 	list := &corev1.ConfigMapList{}
 	if err := params.Client.List(ctx, list, opts...); err != nil {
 		return fmt.Errorf("failed to list: %w", err)
