+++ pkg/collector/reconcile/configmap_test.go
@@ -1,0 +1,141 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package reconcile
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	v1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/tools/record"
+
+	"github.com/open-telemetry/opentelemetry-operator/api/v1alpha1"
+	"github.com/open-telemetry/opentelemetry-operator/internal/config"
+)
+
+func TestDesiredConfigMap(t *testing.T) {
+	t.Run("should return expected config map", func(t *testing.T) {
+		expectedLables := map[string]string{
+			"app.kubernetes.io/managed-by": "opentelemetry-operator",
+			"app.kubernetes.io/instance":   "default.test",
+			"app.kubernetes.io/part-of":    "opentelemetry",
+			"app.kubernetes.io/component":  "opentelemetry-collector",
+			"app.kubernetes.io/name":       "test-collector",
+		}
+
+		expectedData := map[string]string{
+			"collector.yaml": `
+    receivers:
+      jaeger:
+        protocols:
+          grpc:
+    processors:
+
+    exporters:
+      logging:
+
+    service:
+      pipelines:
+        traces:
+          receivers: [jaeger]
+          processors: []
+          exporters: [logging]
+
+`,
+		}
+
+		actual := desiredConfigMap(context.Background(), params())
+
+		assert.Equal(t, "test-collector", actual.Name)
+		assert.Equal(t, expectedLables, actual.Labels)
+		assert.Equal(t, expectedData, actual.Data)
+
+	})
+
+}
+
+func TestExpectedConfigMap(t *testing.T) {
+	t.Run("should create config map", func(t *testing.T) {
+		err := expectedConfigMaps(context.Background(), params(), []v1.ConfigMap{desiredConfigMap(context.Background(), params())}, true)
+		assert.NoError(t, err)
+
+		exists, err := populateObjectIfExists(t, &v1.ConfigMap{}, types.NamespacedName{Namespace: "default", Name: "test-collector"})
+
+		assert.NoError(t, err)
+		assert.True(t, exists)
+	})
+
+	t.Run("should update config map", func(t *testing.T) {
+
+		param := Params{
+			Config: config.New(),
+			Client: k8sClient,
+			Instance: v1alpha1.OpenTelemetryCollector{
+				TypeMeta: metav1.TypeMeta{
+					Kind:       "opentelemetry.io",
+					APIVersion: "v1",
+				},
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test",
+					Namespace: "default",
+					UID:       instanceUID,
+				},
+			},
+			Scheme:   testScheme,
+			Log:      logger,
+			Recorder: record.NewFakeRecorder(10),
+		}
+		cm := desiredConfigMap(context.Background(), param)
+		createObjectIfNotExists(t, "test-collector", &cm)
+
+		err := expectedConfigMaps(context.Background(), params(), []v1.ConfigMap{desiredConfigMap(context.Background(), params())}, true)
+		assert.NoError(t, err)
+
+		actual := v1.ConfigMap{}
+		exists, err := populateObjectIfExists(t, &actual, types.NamespacedName{Namespace: "default", Name: "test-collector"})
+
+		assert.NoError(t, err)
+		assert.True(t, exists)
+		assert.Equal(t, instanceUID, actual.OwnerReferences[0].UID)
+		assert.Equal(t, params().Instance.Spec.Config, actual.Data["collector.yaml"])
+	})
+
+	t.Run("should delete config map", func(t *testing.T) {
+
+		deletecm := v1.ConfigMap{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "test-delete-collector",
+				Namespace: "default",
+				Labels: map[string]string{
+					"app.kubernetes.io/instance":   "default.test",
+					"app.kubernetes.io/managed-by": "opentelemetry-operator",
+				},
+			},
+		}
+		createObjectIfNotExists(t, "test-delete-collector", &deletecm)
+
+		exists, _ := populateObjectIfExists(t, &v1.ConfigMap{}, types.NamespacedName{Namespace: "default", Name: "test-delete-collector"})
+		assert.True(t, exists)
+
+		err := deleteConfigMaps(context.Background(), params(), []v1.ConfigMap{desiredConfigMap(context.Background(), params())})
+		assert.NoError(t, err)
+
+		exists, _ = populateObjectIfExists(t, &v1.ConfigMap{}, types.NamespacedName{Namespace: "default", Name: "test-delete-collector"})
+		assert.False(t, exists)
+	})
+}
