+++ pkg/collector/reconcile/daemonset.go
