+++ pkg/targetallocator/reconcile/suite_test.go
@@ -31,7 +31,6 @@ import (
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/apimachinery/pkg/util/uuid"
 	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/tools/record"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/envtest"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
@@ -49,7 +48,7 @@ var instanceUID = uuid.NewUUID()
 
 func TestMain(m *testing.M) {
 	testEnv = &envtest.Environment{
-		CRDDirectoryPaths: []string{filepath.Join("..", "..", "config", "crd", "bases")},
+		CRDDirectoryPaths: []string{filepath.Join("..", "..", "..", "config", "crd", "bases")},
 	}
 
 	cfg, err := testEnv.Start()
@@ -81,7 +80,7 @@ func TestMain(m *testing.M) {
 	os.Exit(code)
 }
 
-func paramsTA() Params {
+func params() Params {
 	replicas := int32(1)
 	configYAML, err := ioutil.ReadFile("suite_test.yaml")
 	if err != nil {
@@ -118,61 +117,8 @@ func paramsTA() Params {
 				Config:   string(configYAML),
 			},
 		},
-		Scheme:   testScheme,
-		Log:      logger,
-		Recorder: record.NewFakeRecorder(10),
-	}
-}
-
-func paramsCollector() Params {
-	replicas := int32(2)
-	return Params{
-		Config: config.New(),
-		Client: k8sClient,
-		Instance: v1alpha1.OpenTelemetryCollector{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "opentelemetry.io",
-				APIVersion: "v1",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "test",
-				Namespace: "default",
-				UID:       instanceUID,
-			},
-			Spec: v1alpha1.OpenTelemetryCollectorSpec{
-				Ports: []v1.ServicePort{{
-					Name: "web",
-					Port: 80,
-					TargetPort: intstr.IntOrString{
-						Type:   intstr.Int,
-						IntVal: 80,
-					},
-					NodePort: 0,
-				}},
-				Replicas: &replicas,
-				Config: `
-    receivers:
-      jaeger:
-        protocols:
-          grpc:
-    processors:
-
-    exporters:
-      logging:
-
-    service:
-      pipelines:
-        traces:
-          receivers: [jaeger]
-          processors: []
-          exporters: [logging]
-
-`,
-			},
-		},
-		Scheme:   testScheme,
-		Log:      logger,
-		Recorder: record.NewFakeRecorder(10),
+		Scheme: testScheme,
+		Log:    logger,
 	}
 }
